Rework `sortedClassNames` to avoid gotcha
The original `sortedClassNames` would have required fellows to implement
a solution in only one particular way, or else they would hit a gotcha
involving JavaScript `Array.sort()` when sorting an array of objects
by a string property. This will not work in the normal way and is not
covered in the curriculum.

To avoid this, I've reworked the problem to sort by class size instead of
class name. I've also updated the test assertions and renamed the
problem and test files accordingly.