Simplify `mostCommonLetter` requirements
Turning down the difficulty on this one so it is closer in difficulty
level to problems we see on the first assessment. Specifically, I've
removed the requirement around stripping or ignoring non-letter characters
and handling upper vs lower cased characters. In my opinion there is
enough going on here with counting and finding the max to be a fair
challenge for the module 1 assessment.